Reorg waybar conf; icon in conf not scr; +icon wid list README; +window widget
Reorganizing waybar-config.jsonc with widgets listed in the order they appear
left to right in the Waybar. Adding icons used in Waybar to README list of
widgets. Also making it so that widget icons are included in
waybar-config.jsonc not in the shell scripts they call. Adding hyprland/window
widget.